updated
I was unaware that the "sort" option was not active.  You only see this using Linux, and I use windows.
It was buggy and still in development from years ago!  I finished it and activated it.
The "find similar files" algorithm now ignores the destination file that will be overwritten by the source file.
The filename is hightlighted red, so that is obvious anyway.